+++ client/src/services/user-storage.ts
@@ -1,0 +1,6 @@
+import { UserStorageService } from '../application/ports';
+import { useAuthContext } from './store';
+
+export const useUserStorage = (): UserStorageService => {
+  return useAuthContext() as UserStorageService;
+};
